combine api methods into single file pair
