--- backend/services/delete_api_key.go
@@ -1,50 +1,0 @@
-// Copyright 2020, Verizon Media
-// Licensed under the terms of the MIT. See LICENSE file in project root for terms.
-
-package services
-
-import (
-	"context"
-
-	"github.com/theparanoids/ashirt-server/backend"
-	"github.com/theparanoids/ashirt-server/backend/database"
-	"github.com/theparanoids/ashirt-server/backend/policy"
-	"github.com/theparanoids/ashirt-server/backend/server/middleware"
-
-	sq "github.com/Masterminds/squirrel"
-)
-
-type DeleteAPIKeyInput struct {
-	AccessKey string
-	UserSlug  string
-}
-
-func DeleteAPIKey(ctx context.Context, db *database.Connection, i DeleteAPIKeyInput) error {
-	var userID int64
-	var err error
-
-	if userID, err = SelfOrSlugToUserID(ctx, db, i.UserSlug); err != nil {
-		return backend.WrapError("Unable to delete API Key", backend.DatabaseErr(err))
-	}
-
-	if err := policy.Require(middleware.Policy(ctx), policy.CanModifyAPIKeys{UserID: userID}); err != nil {
-		return backend.WrapError("Unwilling to delete API Key", backend.UnauthorizedWriteErr(err))
-	}
-
-	var apiKeyID int64
-
-	err = db.WithTx(ctx, func(tx *database.Transactable) {
-		tx.Get(&apiKeyID, sq.Select("id").
-			From("api_keys").
-			Where(sq.Eq{"user_id": userID, "access_key": i.AccessKey}))
-		tx.Delete(sq.Delete("api_keys").Where(sq.Eq{"id": apiKeyID}))
-	})
-	if err != nil {
-		if database.IsEmptyResultSetError(err) {
-			return backend.WrapError("API key does not exist", backend.UnauthorizedWriteErr(err))
-		}
-		return backend.WrapError("Cannot delete API key", backend.DatabaseErr(err))
-	}
-
-	return nil
-}
